Derive Macro stronghold_derive::GuardDebug

#[derive(GuardDebug)]

A version of the derive Debug trait that blocks parsing the data inside of a struct or enum. Use GuardDebug to block reading and inspection of a data structure via the Debug trait.